[HUMAN App] refactor: separated components form welcome page (#3160)
