+++ packages/apps/human-app/frontend/src/modules/homepage/components/index.ts
@@ -1,3 +1,1 @@
 export * from './home-container';
-export * from './welcome';
-export * from './choose-sign-up-account-type';
